+++ modules/__init__.py
@@ -3,3 +3,4 @@
 from modules.banner import *
 from modules.database import DataBase
 from modules.menu import Menu
+from modules.main import Manager