+++ CHANGELOG.md
@@ -19,3 +19,18 @@ This is the initial version.
 ### Removed
 - json master password system.
 - key system
+
+## [V1.2] - 2022-01-23
+### Added
+- Exceptions (KeyboardInterrupt) and exit config
+- Colors
+- See only one information (`stored_passwords()`)
+    - It shows just the platofrm and the ID
+- Checkmark and x mark
+
+### Changed
+- Manager class in `__main__.py` to `modules/main.py`
+- to look up the password, need to know the ID using the function `stored_password()`
+
+### Removed
+- Old show informations function (`see_all()`)